requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
chromadb>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
